Refactoring fixes
